test(lockdown): move unit tests under __tests__/, add audio-gate snip
Aligns with the repo's dominant pattern (__tests__ folder per feature dir
rather than colocated). Adds a snip that requests the audio format under
lockdown on a cached URL — succeeds only if the audio gate fires, since
firecrawl.dev would otherwise raise AudioUnsupportedUrlError.
