+++ apps/api/src/scraper/scrapeURL/transformers/__tests__/audio.test.ts
@@ -1,4 +1,4 @@
-import { fetchAudio } from "./audio";
+import { fetchAudio } from "../audio";
 
 describe("fetchAudio lockdown guard", () => {
   const originalFetch = global.fetch;
